+++ packages/component-dashboard/client/components/DashboardContent.js
