+++ packages/component-dashboard/client/components/ManuscriptsQuery.js
