+++ packages/component-dashboard/client/components/NewSubmissionButton.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { Mutation } from 'react-apollo'
 import { Button } from '@pubsweet/ui'
 
-import { CREATE_MANUSCRIPT } from './graphql/mutations'
+import { CREATE_MANUSCRIPT } from '../graphql/mutations'
 
 const NewSubmissionButton = ({ history, dataTestId, ...props }) => (
   <Mutation mutation={CREATE_MANUSCRIPT}>
